requirements.txt
python-dotenv>=1.0.0
# 核心依賴
openai-whisper>=20231117
google-generativeai>=0.3.0
yt-dlp>=2023.12.30
requests>=2.31.0
m3u8>=4.0.0
selenium>=4.15.0
tqdm>=4.66.0

# 音訊處理
ffmpeg-python>=0.2.0

# 其他工具
pathlib2>=2.3.7 